requirements.txt
asgiref==3.3.4
astroid==2.5.6
attrs==21.2.0
autopep8==1.5.7
Django==3.2.3
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.1.0
pylint==2.8.2
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
rope==0.19.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
